reorganization and fix geojson rendering
