+++ geostore/db/managers.py
