+++ geostore/models.py
@@ -3,14 +3,11 @@ import json
 import logging
 import os
 import uuid
-from copy import deepcopy
-from functools import reduce
 from itertools import islice
 from tempfile import TemporaryDirectory
 
 import fiona
 import fiona.transform
-from deepmerge import always_merger
 from django.contrib.auth.models import Group
 from django.contrib.gis.db import models
 from django.contrib.gis.db.models.aggregates import Extent
@@ -30,9 +27,9 @@ from fiona.crs import from_epsg
 from mercantile import tiles
 
 from . import GeometryTypes, settings as app_settings
-from .db.mixins import BaseUpdatableModel
+from .db.managers import FeatureQuerySet
+from .db.mixins import BaseUpdatableModel, LayerBasedModelMixin
 from .helpers import ChunkIterator, make_zipfile_bytesio
-from .managers import FeatureQuerySet
 from .routing.decorators import topology_update
 from .signals import save_feature, save_layer_relation
 from .tiles.decorators import zoom_update
@@ -47,108 +44,6 @@ ACCEPTED_PROJECTIONS = [
     'urn:ogc:def:crs:OGC:1.3:CRS84',
     'EPSG:4326',
 ]
-
-
-class LayerBasedModelMixin(BaseUpdatableModel):
-    # Settings schema
-    SETTINGS_DEFAULT = {
-        'metadata': {
-            'attribution': None,  # Json, eg. {'name': 'OSM contributors', href='http://openstreetmap.org'}
-            'licence': None,  # String, eg. 'ODbL'
-            'description': None,  # String
-        },
-        # Tilesets attributes
-        'tiles': {
-            'minzoom': 0,
-            'maxzoom': 22,
-            'pixel_buffer': 4,
-            'features_filter': None,  # Json
-            'properties_filter': None,  # Array of string
-            'features_limit': 10000,
-        }
-    }
-    settings = JSONField(default=dict, blank=True)
-    geom_type = models.IntegerField(choices=GeometryTypes.choices(), null=True)
-
-    @property
-    def is_point(self):
-        return self.layer_geometry in (GeometryTypes.Point,
-                                       GeometryTypes.MultiPoint)
-
-    @property
-    def is_linestring(self):
-        return self.layer_geometry in (GeometryTypes.LineString,
-                                       GeometryTypes.MultiLineString)
-
-    @property
-    def is_polygon(self):
-        return self.layer_geometry in (GeometryTypes.Polygon,
-                                       GeometryTypes.MultiPolygon)
-
-    @property
-    def is_multi(self):
-        return self.layer_geometry in (GeometryTypes.MultiPoint,
-                                       GeometryTypes.MultiLineString,
-                                       GeometryTypes.MultiPolygon)
-
-    @cached_property
-    def layer_geometry(self):
-        """
-        Return the geometry type of the layer using the first feature in
-        the layer if the layer have no geom_type or the geom_type of the layer
-        """
-        if self.geom_type is None:
-            feature = self.features.first()
-            if feature:
-                return feature.geom.geom_typeid
-        return self.geom_type
-
-    @cached_property
-    def settings_with_default(self):
-        return always_merger.merge(deepcopy(self.SETTINGS_DEFAULT), self.settings)
-
-    def layer_settings(self, *json_path):
-        ''' Return the nested value of settings at path json_path.
-            Raise an KeyError if not defined.
-        '''
-        # Dives into settings using args
-        return reduce(
-            lambda a, v: a[v],  # Let raise KeyError on missing key
-            json_path,
-            self.settings) if self.settings is not None else None
-
-    def layer_settings_with_default(self, *json_path):
-        ''' Return the nested value of settings with SETTINGS_DEFAULT as
-            fallback at path json_path.
-            Raise an KeyError if not defined.
-        '''
-        # Dives into settings using args
-        return reduce(
-            lambda a, v: a[v],  # Let raise KeyError on missing key
-            json_path,
-            self.settings_with_default)
-
-    def set_layer_settings(self, *json_path_value):
-        '''Set last parameter as value at the path place into settings
-        '''
-        json_path, value = json_path_value[:-1], json_path_value[-1]
-        # Dive into settings until the last key of path,
-        # and set the corresponding value
-        settings = self.settings
-        for key in json_path[:-1]:
-            s = settings.get(key, {})
-            settings[key] = s
-            settings = s
-        settings[json_path[-1]] = value
-
-        try:
-            # Delete the cached property
-            del self.settings_with_default
-        except AttributeError:
-            pass  # Let's continue, cache was not set
-
-    class Meta:
-        abstract = True
 
 
 class Layer(LayerBasedModelMixin):
@@ -400,8 +295,7 @@ class Layer(LayerBasedModelMixin):
             feature = self.features.filter(pk=nearest_feature.pk)
             feature.update(properties=new_feature.get('properties', {}))
             modified |= feature
-        # clean cache of updated features
-        [feature.clean_vect_tile_cache() for feature in modified]
+
         return modified
 
     @cached_property
@@ -622,7 +516,7 @@ post_save.connect(save_feature, sender=Feature)
 
 class LayerRelation(models.Model):
     RELATION_TYPES = (
-        (None, 'Manuelle'),
+        (None, 'Manual'),
         ('intersects', 'Intersects'),
         ('distance', 'Distance'),
     )
